+++ cirq-core/cirq/__init__.py
@@ -367,6 +367,9 @@ from cirq.transformers import (
     two_qubit_matrix_to_diagonal_and_operations,
     two_qubit_matrix_to_operations,
     two_qubit_matrix_to_sqrt_iswap_operations,
+    two_qubit_gate_product_tabulation,
+    TwoQubitGateTabulation,
+    TwoQubitGateTabulationResult,
     unroll_circuit_op,
     unroll_circuit_op_greedy_earliest,
     unroll_circuit_op_greedy_frontier,
