+++ cirq-core/cirq/protocols/json_test_data/spec.py
@@ -70,6 +70,7 @@ TestSpec = ModuleJsonTestSpec(
         'ThreeQubitDiagonalGate',
         'Timestamp',
         'TwoQubitDiagonalGate',
+        'TwoQubitGateTabulationResult',
         'UnitSweep',
         'StateVectorSimulatorState',
         'StateVectorTrialResult',
