+++ cirq-core/cirq/transformers/heuristic_decompositions/gate_tabulation_math_utils.py
