+++ cirq-google/cirq_google/optimizers/optimize_for_sycamore.py
@@ -23,8 +23,6 @@ from cirq_google.optimizers import (
     convert_to_xmon_gates,
     ConvertToSycamoreGates,
     ConvertToSqrtIswapGates,
-    gate_product_tabulation,
-    GateTabulation,
 )
 
 if TYPE_CHECKING:
@@ -40,7 +38,7 @@ def _get_common_cleanup_optimizers(tolerance: float) -> List[Callable[[cirq.Circ
 
 
 def _get_xmon_optimizers(
-    tolerance: float, tabulation: Optional[GateTabulation]
+    tolerance: float, tabulation: Optional[cirq.TwoQubitGateTabulation]
 ) -> List[Callable[[cirq.Circuit], None]]:
     if tabulation is not None:
         # coverage: ignore
@@ -55,7 +53,7 @@ def _get_xmon_optimizers(
 
 
 def _get_xmon_optimizers_part_cz(
-    tolerance: float, tabulation: Optional[GateTabulation]
+    tolerance: float, tabulation: Optional[cirq.TwoQubitGateTabulation]
 ) -> List[Callable[[cirq.Circuit], None]]:
     if tabulation is not None:
         # coverage: ignore
@@ -69,7 +67,7 @@ def _get_xmon_optimizers_part_cz(
 
 
 def _get_sycamore_optimizers(
-    tolerance: float, tabulation: Optional[GateTabulation]
+    tolerance: float, tabulation: Optional[cirq.TwoQubitGateTabulation]
 ) -> List[Callable[[cirq.Circuit], None]]:
     return [
         ConvertToSycamoreGates(tabulation=tabulation).optimize_circuit,
@@ -79,7 +77,7 @@ def _get_sycamore_optimizers(
 
 
 def _get_sqrt_iswap_optimizers(
-    tolerance: float, tabulation: Optional[GateTabulation]
+    tolerance: float, tabulation: Optional[cirq.TwoQubitGateTabulation]
 ) -> List[Callable[[cirq.Circuit], None]]:
     if tabulation is not None:
         # coverage: ignore
@@ -102,14 +100,14 @@ _OPTIMIZER_TYPES = {
 @lru_cache()
 def _gate_product_tabulation_cached(
     optimizer_type: str, tabulation_resolution: float
-) -> GateTabulation:
+) -> cirq.TwoQubitGateTabulation:
     random_state = np.random.RandomState(51)
     if optimizer_type == 'sycamore':
-        return gate_product_tabulation(
+        return cirq.two_qubit_gate_product_tabulation(
             cirq.unitary(cg_ops.SYC), tabulation_resolution, random_state=random_state
         )
     else:
-        raise NotImplementedError(f"Gate tabulation not supported for {optimizer_type}")
+        raise NotImplementedError(f"Two qubit gate tabulation not supported for {optimizer_type}")
 
 
 def optimized_for_sycamore(
@@ -155,7 +153,7 @@ def optimized_for_sycamore(
             f'types are: {_OPTIMIZER_TYPES.keys()}'
         )
 
-    tabulation: Optional[GateTabulation] = None
+    tabulation: Optional[cirq.TwoQubitGateTabulation] = None
     if tabulation_resolution is not None:
         tabulation = _gate_product_tabulation_cached(optimizer_type, tabulation_resolution)
