+++ cirq-google/cirq_google/optimizers/two_qubit_gates/example.py
@@ -24,13 +24,14 @@ import numpy as np
 from matplotlib import pyplot as plt
 
 import cirq
-from cirq_google.optimizers.two_qubit_gates.gate_compilation import gate_product_tabulation
-from cirq_google.optimizers.two_qubit_gates.math_utils import unitary_entanglement_fidelity
+from cirq.transformers.heuristic_decompositions.gate_tabulation_math_utils import (
+    unitary_entanglement_fidelity,
+)
 from cirq.testing import random_special_unitary
 
 
 def main(samples: int = 1000, max_infidelity: float = 0.01):
-    """Demonstration of the usage of the GateTabulation gate compiler.
+    """Demonstration of the usage of the TwoQubitGateTabulation gate compiler.
 
     Args:
         samples: Number of random 2-qubit unitary samples to compile.
@@ -43,14 +44,14 @@ def main(samples: int = 1000, max_infidelity: float = 0.01):
     phi = np.pi / 24
     base = cirq.unitary(cirq.FSimGate(theta, phi))
 
-    # The GateTabulation object is essentially a tabulation of many randomly
+    # The TwoQubitGateTabulation object is essentially a tabulation of many randomly
     # generated gate products (of the form A k A or A k A k A), along with their
     # associate KAK vectors. The parameter max_infidelity determines the
     # approximate "density" of the tabulated gates. Specifically, it bounds the
     # typical distance between an arbitrary two-qubit gate and the nearest
     # tabulated gate.
     start = time()
-    tabulation = gate_product_tabulation(base, max_infidelity)
+    tabulation = cirq.two_qubit_gate_product_tabulation(base, max_infidelity)
 
     print(tabulation.summary)
     print(f'Gate tabulation time : {time() - start} seconds.')
